split cards up into a series of components starting at available or registered points
