+++ src/components/courses/RegisteredCard.vue
@@ -1,0 +1,121 @@
+<template>
+  <v-card v-if="course" outlined class="course-card">
+    <v-chip class="status" small :ripple="false" :color="statusColor">{{ status }}</v-chip>
+    <v-row>
+      <v-col>
+        <CourseTitle :course="course" />
+      </v-col>
+      <v-col>
+        <CourseTimes :course="course" :timezone="user.timezone" />
+      </v-col>
+      <v-col>
+        <CourseTeacher :course="course" />
+      </v-col>
+    </v-row>
+    <v-list-item>
+      <v-list-item-subtitle>
+        <CourseStudents :registered="registeredForCourse" />
+        <CoursePaymentDue
+          v-if="status === 'paid'"
+          :totalDue="totalDue"
+          :totalDueBy="totalDueBy"
+          :timezone="user.timezone"
+        />
+      </v-list-item-subtitle>
+    </v-list-item>
+    <v-card-actions>
+      <button @click="testClicky(registeredCourseId)">test</button>
+    </v-card-actions>
+  </v-card>
+</template>
+
+<script>
+import CourseService from "@/service/courseService";
+import CourseTitle from "@/components/courses/card/CourseTitle";
+import CourseTimes from "@/components/courses/card/CourseTimes";
+import CourseTeacher from "@/components/courses/card/CourseTeacher";
+import CourseStudents from "@/components/courses/card/CourseStudents";
+import CoursePaymentDue from "@/components/courses/card/CoursePaymentDue";
+import { formatDateToLocal, getWeeks } from "@/utils/timeUtils";
+export default {
+  name: "RegisteredCard",
+  components: {
+    CourseTitle,
+    CourseTimes,
+    CourseTeacher,
+    CourseStudents,
+    CoursePaymentDue
+  },
+  props: {
+    course: {
+      type: Object,
+      default: null
+    },
+    user: {
+      type: Object,
+      default: null
+    },
+    family: {
+      type: Object,
+      default: null
+    },
+    type: {
+      type: String,
+      default: null
+    },
+    registeredIds: {
+      type: Array,
+      default: () => []
+    },
+    registeredCourseId: {
+      type: Number,
+      default: null
+    },
+    status: {
+      type: String,
+      default: null
+    },
+    totalDue: {
+      type: Number,
+      default: null
+    },
+    totalDueBy: {
+      type: String,
+      default: null
+    }
+  },
+  methods: {
+    formatDateToLocal,
+    getWeeks,
+    testClicky(id) {
+      CourseService.fetchRegisteredCourseSessions(id);
+    }
+  },
+  computed: {
+    registeredForCourse() {
+      const regArr = [];
+      if (this.registeredIds[0])
+        regArr.push(this.family[this.registeredIds[0]]);
+      if (this.registeredIds[1])
+        regArr.push(this.family[this.registeredIds[1]]);
+      if (this.registeredIds[2])
+        regArr.push(this.family[this.registeredIds[2]]);
+      return regArr;
+    },
+    statusColor() {
+      if (this.$props.status === "paid") return "primary";
+      return "secondary";
+    }
+  }
+};
+</script>
+<style lang="scss" scoped>
+.course-card {
+  padding-top: 2rem;
+}
+.status {
+  position: absolute;
+  right: 0.5rem;
+  top: 0.5rem;
+}
+</style>
