+++ src/components/courses/card/CourseTitle.vue
@@ -1,0 +1,23 @@
+<template>
+  <v-list-item three-line>
+    <v-list-item-icon>
+      <v-icon color="primary">mdi-school</v-icon>
+    </v-list-item-icon>
+    <v-list-item-content v-if="course.specialty">
+      <div class="overline">{{ course.specialty.subject }}</div>
+      <v-list-item-title class="headline mb-1">{{ course.specialty.category }}</v-list-item-title>
+      <v-list-item-subtitle>{{ course.specialty.focus_areas }}</v-list-item-subtitle>
+    </v-list-item-content>
+  </v-list-item>
+</template>
+
+<script>
+export default {
+  props: {
+    course: {
+      type: Object,
+      default: null
+    }
+  }
+};
+</script>
