+++ src/service/courseService.js
@@ -1,16 +1,9 @@
 import absAPI from "@/api/absAPI";
 import authHeader from "./authHeader";
 
-class UserService {
+class CourseService {
   async fetchAvailableCourses(email) {
     const res = await absAPI.get("/courses", {
-      params: { user_email: email },
-      headers: authHeader()
-    });
-    return res.data;
-  }
-  async fetchRegisteredCourses(email) {
-    const res = await absAPI.get("/registrations", {
       params: { user_email: email },
       headers: authHeader()
     });
@@ -22,6 +15,20 @@ class UserService {
     });
     return res.data;
   }
+  async fetchRegisteredCourses(email) {
+    const res = await absAPI.get("/registrations", {
+      params: { user_email: email },
+      headers: authHeader()
+    });
+    return res.data;
+  }
+  async fetchRegisteredCourseSessions(id) {
+    const res = await absAPI.get(`/registrations/${id}/class_sessions`, {
+      params: { id: id },
+      headers: authHeader()
+    });
+    return res.data;
+  }
 }
 
-export default new UserService();
+export default new CourseService();
