+++ src/views/Courses.vue
@@ -4,32 +4,21 @@
       <v-card-text>
         <v-container>
           <v-row>
-            <v-col
-              v-if="hasRegisteredCourses > 0"
-              cols="12"
-              class="courses courses-registered"
-            >
+            <v-col v-if="hasRegisteredCourses > 0" cols="12" class="courses courses-registered">
               <strong class="title">Registered Courses</strong>
               <p>Courses with registered students.</p>
               <v-row v-if="!loadingRegisteredCoursesFetch">
-                <v-col
-                  v-for="item in registeredCourses"
-                  :key="item.id"
-                  cols="12"
-                  md="6"
-                >
-                  <CourseCard
-                    type="registered"
+                <v-col v-for="item in registeredCourses" :key="item.id" cols="12">
+                  <RegisteredCard
                     :course="item.course"
                     :user="activeUser"
                     :family="currentFamilyMembers"
-                    :hasFamilyMembers="hasFamilyMembers"
-                    :loadingFamily="loadingFamilyMemberFetch"
                     :registeredIds="[
                       item.primary_family_member_id,
                       item.secondary_family_member_id,
                       item.tertiary_family_member_id
                     ]"
+                    :registeredCourseId="item.id"
                     :status="item.status"
                     :totalDue="item.total_due / 100"
                     :totalDueBy="item.total_due_by"
@@ -40,18 +29,12 @@
             <v-col cols="12" class="courses courses-available">
               <strong class="title">Courses Available</strong>
               <p>View available courses and register students.</p>
-              <p v-if="hasCourses <= 0 && !loadingCoursesFetch">
-                No new courses are available at this time.
-              </p>
+              <p
+                v-if="hasCourses <= 0 && !loadingCoursesFetch"
+              >No new courses are available at this time.</p>
               <v-row v-else-if="hasCourses > 0 && !loadingCoursesFetch">
-                <v-col
-                  v-for="course in availableCourses"
-                  :key="course.id"
-                  cols="12"
-                  md="6"
-                >
-                  <CourseCard
-                    type="available"
+                <v-col v-for="course in availableCourses" :key="course.id" cols="12" md="6">
+                  <AvailableCard
                     :course="course"
                     :user="activeUser"
                     :family="currentFamilyMembers"
@@ -71,7 +54,8 @@
 <script>
 import { mapGetters } from "vuex";
 import DashboardWrap from "@/components/layouts/DashboardWrap";
-import CourseCard from "@/components/courses/CourseCard.vue";
+import AvailableCard from "@/components/courses/AvailableCard.vue";
+import RegisteredCard from "@/components/courses/RegisteredCard.vue";
 import actionTypes from "@/store/actions";
 const { courses, family } = actionTypes;
 export default {
@@ -79,7 +63,7 @@ export default {
   metaInfo: {
     title: "Register for available courses"
   },
-  components: { DashboardWrap, CourseCard },
+  components: { DashboardWrap, AvailableCard, RegisteredCard },
   async created() {
     await this.$store.dispatch(family.request, this.activeUser.email);
     await this.$store.dispatch(courses.request, this.activeUser.email);
@@ -104,4 +88,14 @@ export default {
 };
 </script>
 
-<style lang="scss" scoped></style>
+<style lang="scss" scoped>
+.v-card.course-card {
+  border-color: $brand-blue;
+}
+.course-card-footer {
+  background: $white-smoke;
+}
+.status {
+  width: 100%;
+}
+</style>
